Refactor auth and session persistence out of sync UI classes
- Moved `saveUserInfoPref` to `UserSessionManager` to consolidate authentication and session logic into a single boundary.
- Removed duplicated `saveUserInfoPref` logic from `ProcessUserDataActivity`.
- Updated `SyncActivity`, `LoginActivity`, and `GuestLoginExtensions` to use the injected `UserSessionManager` for writing user credentials.
- Removed direct `DatabaseService` injection and explicit Dispatchers from `SyncActivity`, deferring instead to `UserRepository`. Child activities like `DashboardElementActivity` and `DictionaryActivity` now explicitly inject `DatabaseService` where required.
