+++ app/src/main/java/org/ole/planet/myplanet/services/UserSessionManager.kt
@@ -10,9 +10,11 @@ import kotlinx.coroutines.withContext
 import org.ole.planet.myplanet.di.ApplicationScope
 import org.ole.planet.myplanet.model.RealmMyLibrary
 import org.ole.planet.myplanet.model.RealmUser
+import android.content.SharedPreferences
 import org.ole.planet.myplanet.repository.ActivitiesRepository
 import org.ole.planet.myplanet.repository.UserRepository
 import org.ole.planet.myplanet.utils.DispatcherProvider
+import org.ole.planet.myplanet.utils.SecurePrefs
 
 class UserSessionManager @Inject constructor(
     @ApplicationContext private val context: Context,
@@ -34,6 +36,23 @@ class UserSessionManager @Inject constructor(
 
     suspend fun getUserModel(): RealmUser? {
         return userRepository.getUserModelSuspending()
+    }
+
+    suspend fun saveUserInfoPref(settings: SharedPreferences, password: String?, user: RealmUser?) {
+        withContext(dispatcherProvider.io) {
+            SecurePrefs.saveCredentials(context, settings, user?.name, password)
+        }
+        sharedPrefManager.setUserId(user?.id ?: "")
+        sharedPrefManager.setUserName(user?.name ?: "")
+        sharedPrefManager.rawPreferences.edit().apply {
+            remove("password")
+            putString("firstName", user?.firstName)
+            putString("lastName", user?.lastName)
+            putString("middleName", user?.middleName)
+            user?.userAdmin?.let { putBoolean("isUserAdmin", it) }
+            putLong("lastLogin", System.currentTimeMillis())
+            apply()
+        }
     }
 
     fun onLogin() {
