+++ app/src/main/java/org/ole/planet/myplanet/ui/sync/GuestLoginExtensions.kt
@@ -65,7 +65,7 @@ fun LoginActivity.showGuestLoginDialog(userRepository: UserRepository) {
                         toast(this@showGuestLoginDialog, getString(R.string.unable_to_login))
                     } else {
                         saveUsers(username, "", "guest")
-                        saveUserInfoPref(settings, "", model)
+                        profileDbHandler.saveUserInfoPref(settings, "", model)
                         onLogin()
                     }
                 }
